Misc: Moving the Angular gallery app to `packages/angular`
