+++ packages/angular/gallery/src/index.html
@@ -4,7 +4,7 @@
   <head>
     <meta http-equiv="content-type" content="text/html; charset=utf-8" />
     <title>Angular Gallery Viewer</title>
-    <link rel="shortcut icon" href="#" />
+    <link rel="shortcut icon" href="#" />      
   </head>
   <style>
     body {
@@ -17,6 +17,5 @@
   </style>
   <body>
     <app-component></app-component>
-    <script type="text/javascript" src="dist/angular.index.js"></script>
   </body>
 </html>
