+++ packages/angular/licences.txt
@@ -1,25 +1,30 @@
-name                           license period  license type  installed version  author
-----                           --------------  ------------  -----------------  ------
-tslib                          perpetual       0BSD          2.4.1              Microsoft Corp.
-@angular-devkit/build-angular  perpetual       MIT           12.2.18            Angular Authors
-@angular/cli                   perpetual       MIT           12.2.18            Angular Authors
-@angular/common                perpetual       MIT           12.2.17            angular
-@angular/compiler              perpetual       MIT           12.2.17            angular
-@angular/compiler-cli          perpetual       MIT           12.2.17            n/a
-@angular/core                  perpetual       MIT           12.2.17            angular
-@types/jasmine                 perpetual       MIT           3.6.11             n/a
-@types/node                    perpetual       MIT           12.20.55           n/a
-jasmine-core                   perpetual       MIT           3.7.1              n/a
-karma                          perpetual       MIT           6.3.20             Vojta Jína <vojta.jina@gmail.com>
-karma-chrome-launcher          perpetual       MIT           3.1.1              Vojta Jina <vojta.jina@gmail.com>
-karma-coverage                 perpetual       MIT           2.0.3              SATO taichi <ryushi@gmail.com>
-karma-jasmine                  perpetual       MIT           4.0.2              Vojta Jina <vojta.jina@gmail.com>
-karma-jasmine-html-reporter    perpetual       MIT           1.7.0
-ng-packagr                     perpetual       MIT           12.2.7             David Herges <david@spektrakel.de>
-rollup                         perpetual       MIT           2.75.5             Rich Harris
-rollup-plugin-typescript2      perpetual       MIT           0.31.2             @ezolenko
-ttypescript                    perpetual       MIT           1.5.13             cevek
-@angular/common                perpetual       MIT           12.2.17            angular
-@angular/compiler              perpetual       MIT           12.2.17            angular
-@angular/core                  perpetual       MIT           12.2.17            angular
-@unovis/ts                     perpetual       Apache-2.0    1.2.1              Nikita Rokotyan, F5 Inc. <nikita@f5.com> (https://rokotyan.com)
+name                               license period  license type  installed version  author
+----                               --------------  ------------  -----------------  ------
+tslib                              perpetual       0BSD          2.4.1              Microsoft Corp.
+@angular-devkit/build-angular      perpetual       MIT           12.2.18            Angular Authors
+@angular/cli                       perpetual       MIT           12.2.18            Angular Authors
+@angular/common                    perpetual       MIT           12.2.17            angular
+@angular/compiler                  perpetual       MIT           12.2.17            angular
+@angular/compiler-cli              perpetual       MIT           12.2.17            n/a
+@angular/core                      perpetual       MIT           12.2.17            angular
+@angular/platform-browser          perpetual       MIT           12.2.17            angular
+@angular/platform-browser-dynamic  perpetual       MIT           12.2.17            angular
+@types/jasmine                     perpetual       MIT           3.6.11             n/a
+@types/node                        perpetual       MIT           12.20.55           n/a
+jasmine-core                       perpetual       MIT           3.7.1              n/a
+karma                              perpetual       MIT           6.3.20             Vojta Jína <vojta.jina@gmail.com>
+karma-chrome-launcher              perpetual       MIT           3.1.1              Vojta Jina <vojta.jina@gmail.com>
+karma-coverage                     perpetual       MIT           2.0.3              SATO taichi <ryushi@gmail.com>
+karma-jasmine                      perpetual       MIT           4.0.2              Vojta Jina <vojta.jina@gmail.com>
+karma-jasmine-html-reporter        perpetual       MIT           1.7.0
+ng-packagr                         perpetual       MIT           12.2.7             David Herges <david@spektrakel.de>
+rollup                             perpetual       MIT           2.75.5             Rich Harris
+rollup-plugin-typescript2          perpetual       MIT           0.31.2             @ezolenko
+ttypescript                        perpetual       MIT           1.5.13             cevek
+webpack                            perpetual       MIT           5.88.2             Tobias Koppers @sokra
+zone.js                            perpetual       MIT           0.14.0             Brian Ford
+@unovis/shared                     perpetual       Apache-2.0    1.2.1              Nikita Rokotyan, F5 Inc. <nikita@f5.com> (https://rokotyan.com)
+@unovis/ts                         perpetual       Apache-2.0    1.2.1              Nikita Rokotyan, F5 Inc. <nikita@f5.com> (https://rokotyan.com)
+@angular/common                    perpetual       MIT           12.2.17            angular
+@angular/compiler                  perpetual       MIT           12.2.17            angular
+@angular/core                      perpetual       MIT           12.2.17            angular
