+++ packages/angular/package.json
@@ -29,7 +29,8 @@
     "build": "npx ng build",
     "watch": "npx ng build --watch --configuration development",
     "test": "npx ng test",
-    "generate": "rollup -c; node .autogen.cjs; rm .autogen.cjs"
+    "generate": "rollup -c; node .autogen.cjs; rm .autogen.cjs",
+    "gallery": "node gallery"
   },
   "main": "dist/lib/bundles/unovis-angular.umd.js",
   "module": "dist/lib/esm2015/public-api.js",
@@ -43,10 +44,11 @@
     "tslib": "^2.3.1"
   },
   "peerDependencies": {
+    "@unovis/shared": "*",
+    "@unovis/ts": "1.2.1",
     "@angular/common": "12 - 16",
     "@angular/compiler": "12 - 16",
-    "@angular/core": "12 - 16",
-    "@unovis/ts": "1.2.1"
+    "@angular/core": "12 - 16"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^12.0.3",
@@ -55,6 +57,8 @@
     "@angular/compiler": "^12.0.3",
     "@angular/compiler-cli": "^12.0.3",
     "@angular/core": "^12.0.3",
+    "@angular/platform-browser": "^12.0.3",
+    "@angular/platform-browser-dynamic": "^12.0.3",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.7.0",
@@ -66,7 +70,9 @@
     "ng-packagr": "^12.0.0",
     "rollup": "^2.61.1",
     "rollup-plugin-typescript2": "^0.31.1",
-    "ttypescript": "^1.5.13"
+    "ttypescript": "^1.5.13",
+    "webpack": "^5.88.2",
+    "zone.js": "^0.14.0"
   },
   "overrides": {
     "minimatch": "3.0.5",
