+++ packages/shared/examples/free-brush-scatters/free-brush-scatters.component.html
@@ -21,9 +21,7 @@
       [autoHide]="false"
       [mode]="brushMode"
       [onBrushEnd]="setSelection"
-      [selectionMinLength]="[0, 0]"
-      [x]="x"
-      [y]="y">
+      [selectionMinLength]="[0, 0]">
     </vis-free-brush>
   </vis-xy-container>
 </div>
