--- packages/website/tsconfig.angular.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "rootDir": "src/examples",
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true
-  },
-  "files": [
-    "src/examples/angular.index.ts"
-  ],
-  "include": [
-    "src/examples/**/*",
-  ],
-  "exclude": [
-    "node_modules",
-    "*/node_modules/*",
-    "build",
-    "src/**/*.tsx",
-    "docs/**/*.tsx"
-  ]
-}
